[#IP-319] remove e2e compatibility tests for fp-ts 1.x
