+++ e2e/package.json
@@ -21,9 +21,9 @@
   },
   "dependencies": {
     "@pagopa/openapi-codegen-ts": "../",
-    "@pagopa/ts-commons": "^9.1.0",
-    "fp-ts": "1.19.5",
-    "io-ts": "1.8.5"
+    "@pagopa/ts-commons": "^10.0.0",
+    "fp-ts": "^2.10.5",
+    "io-ts": "^2.2.16"
   },
   "jest": {
     "testEnvironment": "node",
