--- e2e_fp-ts-v2/.gitignore
@@ -1,8 +1,0 @@
-node_modules/
-**/node_modules/
-**/generated/
-npm-debug.log
-yarn-error.log
-local*
-.vscode
-package-lock.json
