--- e2e_fp-ts-v2/src/teardown.ts
@@ -1,3 +1,0 @@
-import { stopAllServers } from "./server";
-
-export default stopAllServers;
