--- e2e_fp-ts-v2/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2015",
-    "rootDir": "src",
-    "outDir": "dist",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "strict": true,
-    "noImplicitAny": true,
-    "declaration": true,
-    "sourceMap": true,
-    "lib": ["es2015", "dom"]
-  },
-  "include": ["src/**/*", "generated/**/*"],
-  "exclude": ["node_modules", "src/**/__tests__/*"]
-}
